docs: document every parameter, and move the gadget work out of cli.py
Once the sections are numpy-shaped, pydocstyle can check them, and D417
found real gaps: inject_gadget_into_apk documented three of its nine
parameters and named one of them wrong (apk, not apk_path), and
build_signer_command and three apk_utils helpers described a parameter
without giving it a type, so the checker could not match it. All of them
are documented now.

The longer docstrings pushed cli.py six lines past pylint's limit. Rather
than trim them, the gadget work moves to scripts/gadget.py, following the
split the other modules already have: choosing and downloading the gadget,
naming it, preparing lib/<abi>, writing the config, and the injection that
ties them together. cli.py is 505 lines and holds option validation and the
click command. tests/test_gadget.py follows.

No executable change: stripping the docstrings and comparing the AST of all
51 top-level definitions against trunk shows none missing and none changed,
and the injection produces identical output for all ten option combinations.
